% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpk.R
\name{bioc_software_packagelist}
\alias{bioc_software_packagelist}
\title{get vector of Bioc software package names}
\usage{
bioc_software_packagelist()
}
\value{
character vector
}
\description{
get vector of Bioc software package names
}
