% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexshi.R
\name{browse_checks}
\alias{browse_checks}
\title{developer report on package check process in shiny}
\usage{
browse_checks(con)
}
\arguments{
\item{con}{SQLiteConnection where tables of check results from rcmdcheck are stored}
}
\description{
developer report on package check process in shiny
}
\note{
dbDisconnect is run if the stop button is pressed to end shiny session.  If
stopped via ctrl-C or error, the connection may remain and require closure.
}
\examples{
if (interactive()) {
 con = RSQLite::dbConnect(RSQLite::SQLite(), 
  system.file("sqlite/demo2.sqlite", package="BiocBBSpack"), flags=RSQLite::SQLITE_RO)
 browse_checks(con)
}
}
