% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build1.R
\name{build1}
\alias{build1}
\title{prepare and build a package tarball}
\usage{
build1(srcpath, dest = ".", ...)
}
\arguments{
\item{srcpath}{character(1) path to source folder for a package}

\item{dest}{character(1) destination folder}
}
\description{
prepare and build a package tarball
}
