% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plus.R
\name{build_plus}
\alias{build_plus}
\title{use rcmdcheck and BiocCheck on a folder with package source checked out}
\usage{
build_plus(folder, error_on = "never")
}
\arguments{
\item{folder}{source folder for R package}

\item{error_on}{character(1) see rcmdcheck::rcmdcheck; default is to proceed when error occurs in a check stage}
}
\description{
use rcmdcheck and BiocCheck on a folder with package source checked out
}
