% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_update_and_pull.R
\name{check_for_update_and_pull}
\alias{check_for_update_and_pull}
\title{check a git repo for activity at remote, specifically modifications to DESCRIPTION}
\usage{
check_for_update_and_pull(repo_path)
}
\arguments{
\item{repo_path}{character(1) path for a git repo}
}
\value{
a vector of results of git diff
}
\description{
check a git repo for activity at remote, specifically modifications to DESCRIPTION
}
\note{
Main purpose is to achieve the side effect of git pull in the 
repo when appropriate.  R session visits repo, performs git commands using
system(), and returns to folder where this function was called.
Probably should add code to report on the difference seen, checking explicitly
for version tag increase.
}
