% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_shiny.R
\name{chk_shiny}
\alias{chk_shiny}
\title{simple interface to details about check}
\usage{
chk_shiny(lis)
}
\arguments{
\item{lis}{output of rcmdcheck for one package}
}
\description{
simple interface to details about check
}
\note{
in development
}
