% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreset.R
\name{coreset}
\alias{coreset}
\title{define a small collection of packages of interest}
\usage{
coreset()
}
\value{
a character vector
}
\description{
define a small collection of packages of interest
}
\examples{
coreset()
}
