% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{covr_tab.coverage}
\alias{covr_tab.coverage}
\title{produce textual summary of covr analysis}
\usage{
\method{covr_tab}{coverage}(x, group = c("filename", "functions"), by = "line", ...)
}
\arguments{
\item{x}{result of `package_coverage`}

\item{group}{character(1)}

\item{by}{character(1)}

\item{\dots}{not used}
}
\description{
produce textual summary of covr analysis
}
\note{
This is because package_coverage uses message to report summaries.
}
