% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{covr_tab}
\alias{covr_tab}
\title{tabulator for covr outputs}
\usage{
\method{covr_tab}{coverage}(x, group = c("functions"), by = "line", ...)
}
\arguments{
\item{x}{instance of coverage S3 class from covr package}

\item{group}{character(1) tells how to group data, see covr:::print.coverage}

\item{by}{character(1) report by line or expression}

\item{\dots}{unused}
}
\description{
tabulator for covr outputs
}
