% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{covr_tab}
\alias{covr_tab}
\alias{covr_tab.coverage}
\title{tabulator for covr outputs}
\usage{
\method{covr_tab}{coverage}(x, group = c("functions"), by = "line", ...)

\method{covr_tab}{coverage}(x, group = c("functions"), by = "line", ...)
}
\arguments{
\item{x}{result of `package_coverage`}

\item{group}{character(1)}

\item{by}{character(1)}

\item{\dots}{not used}
}
\description{
tabulator for covr outputs

produce textual summary of covr analysis
}
\note{
This is because package_coverage uses message to report summaries.  We
use the 'functions' setting for 'group' by default.
}
