% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covtabs_to_df.R
\name{covtabs_to_df}
\alias{covtabs_to_df}
\title{convert result of a collection of `covr_tab.coverage` results to a single data.frame}
\usage{
covtabs_to_df(x)
}
\arguments{
\item{x}{a list of results of `covr_tab.coverage`}
}
\value{
a data.frame instane with columns `package`, `func`, and `pctcov`
}
\description{
convert result of a collection of `covr_tab.coverage` results to a single data.frame
}
