% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endow_ubuntu.R
\name{default_packtxt}
\alias{default_packtxt}
\title{NOT EXPORTED -- use docker, bioconductor/bioconductor_full:(devel or release)
output of sudo apt list --installed for a suitable ubuntu environment}
\usage{
default_packtxt()
}
\value{
a path
}
\description{
NOT EXPORTED -- use docker, bioconductor/bioconductor_full:(devel or release)
output of sudo apt list --installed for a suitable ubuntu environment
}
