% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endow_ubuntu.R
\name{endow_ubuntu}
\alias{endow_ubuntu}
\title{use sudo apt-get in system() to endow and ubuntu system for BBS activities}
\usage{
endow_ubuntu(packtxtfun = default_packtxt)
}
\arguments{
\item{packtxtfun}{a function providing a path to output of sudo apt --list installed}
}
\value{
Value of system().
}
\description{
use sudo apt-get in system() to endow and ubuntu system for BBS activities
}
\note{
Uses sudo apt-get install.  Also, there will be a EULA query and a services query that must be handled interactively.
}
