% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getver.R
\name{getver}
\alias{getver}
\title{in folder superior to a built package, retrieve version}
\usage{
getver(x)
}
\arguments{
\item{x}{character(1) name of folder holding built package}
}
\description{
in folder superior to a built package, retrieve version
}
