% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_shiny.R
\name{ibc}
\alias{ibc}
\title{use rcmdcheck on a package}
\usage{
ibc(folder, error_on = "never")
}
\arguments{
\item{folder}{source folder for R package}

\item{error_on}{character(1) see rcmdcheck::rcmdcheck; default is to proceed when error occurs in a check stage}
}
\description{
use rcmdcheck on a package
}
