% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_tests.R
\name{ingest_tests}
\alias{ingest_tests}
\title{ingest test sources for a package}
\usage{
ingest_tests(src_folder)
}
\arguments{
\item{src_folder}{path to an R package folder tree (i.e., folder that includes DESCRIPTION)}
}
\value{
named list with one element per test file
}
\description{
ingest test sources for a package
}
\note{
looks for files in tests/testthat and inst/unitTests and applies readLines to these
}
