% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_support.R
\name{list_packs_to_update}
\alias{list_packs_to_update}
\title{compare destdir to srcdir and list package paths needing tarballs}
\usage{
list_packs_to_update(srcdir, destdir)
}
\arguments{
\item{srcdir}{folder holding multiple R package source folders}

\item{destdir}{folder where tarballs may be found}
}
\description{
compare destdir to srcdir and list package paths needing tarballs
}
\note{
Elementary filename processing only -- no version checking at this time
}
