% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreset.R
\name{PackageSet}
\alias{PackageSet}
\title{constructor for PackageSet instances}
\usage{
PackageSet(cvec)
}
\arguments{
\item{cvec}{character() vector}
}
\description{
constructor for PackageSet instances
}
\note{
Will issue message if some element of cvec is not
found in BiocPkgTools::biocPkgList() result
}
