% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_support.R
\name{parallel_tarballs}
\alias{parallel_tarballs}
\title{create tarballs from source folders}
\usage{
parallel_tarballs(paths, dest = ".", ncores = 6)
}
\arguments{
\item{paths}{vector of paths to R package source folders}

\item{dest}{location where pkgbuild::build will deliver tarball}

\item{ncores}{sets options(mc.cores) value}
}
\description{
create tarballs from source folders
}
