% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endow_ubuntu.R
\name{provision_ubuntu}
\alias{provision_ubuntu}
\title{NOT EXPORTED -- use docker, bioconductor/bioconductor_full:(devel or release)
use sudo apt-get in system() to provision an ubuntu system for BBS activities}
\usage{
provision_ubuntu(packtxtfun = default_packtxt)
}
\arguments{
\item{packtxtfun}{a function providing a path to output of sudo apt --list installed}
}
\value{
Value of system().
}
\description{
NOT EXPORTED -- use docker, bioconductor/bioconductor_full:(devel or release)
use sudo apt-get in system() to provision an ubuntu system for BBS activities
}
\note{
Uses sudo apt-get install.  Also, there will be a EULA query and a services query that must be handled interactively.
it may be necessary to use commands  sudo add-apt-repository -y ppa:opencpu/poppler;
   sudo apt-get update
   sudo sudo apt-get install -y libpoppler-cpp-dev
}
