% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_local_gits.R
\name{refresh_local_gits}
\alias{refresh_local_gits}
\title{perform git pull for modified packages in a repo collection}
\usage{
refresh_local_gits(gitspath, version = "3.11")
}
\arguments{
\item{gitspath}{character(1) path to repos}

\item{version}{character(1) version tag for Bioconductor packages}
}
\value{
list of folder names visited
}
\description{
perform git pull for modified packages in a repo collection
}
\note{
Simply uses lapply to visit folders and run 'git pull' in system()
}
