% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipit.R
\name{zipit}
\alias{zipit}
\title{tar cz a binary package into 'root' folder}
\usage{
zipit(x, root = "/home/rstudio/zips")
}
\arguments{
\item{x}{character(1) package name}

\item{root}{character(1) folder path where zip will lie}
}
\description{
tar cz a binary package into 'root' folder
}
\note{
package name includes R_x86_64-pc-linux-gnu before .tar.gz
}
