% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endow_ubuntu.R
\name{default_packtxt}
\alias{default_packtxt}
\title{output of sudo apt --list installed for a suitable ubuntu environment}
\usage{
default_packtxt()
}
\value{
a path
}
\description{
output of sudo apt --list installed for a suitable ubuntu environment
}
