% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_support.R
\name{parallel_tarballs}
\alias{parallel_tarballs}
\title{create tarballs from source folders}
\usage{
parallel_tarballs(paths, dest = ".", ncores = 6)
}
\arguments{
\item{paths}{vector of paths to R package source folders}

\item{dest}{location where pkgbuild::build will deliver tarball}

\item{ncores}{sets options(mc.cores) value}
}
\value{
the result of parallel::mclapply, which may include some try-error objects
}
\description{
create tarballs from source folders
}
\note{
This function will call `build1_with_buildsink` to capture the build log messages in a file.  The file will have suffix `.bldlog.txt`.
If the build for package i throws a try-error, that object will be saved in an
object basename(paths[i])_err, serialized to basename(paths[i])_err.rda
}
